services/citizen-api/pyproject.toml
[project]
name = "citizen-api"
version = "0.1.0"
description = "SMWCS Citizen-facing REST API"
authors = [
    {name = "dee-mee",email = "derekmuriuki2@gmail.com"}
]
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.42.0,<0.43.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "motor (>=3.7.1,<4.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "confluent-kafka (>=2.13.2,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.48,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "prometheus-client (>=0.24.1,<0.25.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "ruff (>=0.15.7,<0.16.0)"
]
